refactor: moved telemetry to own package & updated some types for readability
